+++ modules/model/src/main/java/net/pantasystem/milktea/model/user/FollowRequestRepository.kt
@@ -1,0 +1,8 @@
+package net.pantasystem.milktea.model.user
+
+interface FollowRequestRepository {
+
+    suspend fun accept(userId: User.Id) : Boolean
+
+    suspend fun reject(userId: User.Id) : Boolean
+}
