+++ modules/model/src/main/java/net/pantasystem/milktea/model/user/UserRepository.kt
@@ -26,10 +26,6 @@ interface UserRepository {
 
     suspend fun unblock(userId: User.Id): Boolean
 
-    suspend fun acceptFollowRequest(userId: User.Id) : Boolean
-
-    suspend fun rejectFollowRequest(userId: User.Id) : Boolean
-
     suspend fun report(report: Report) : Boolean
 
     suspend fun findUsers(accountId: Long, query: FindUsersQuery): List<User>
